feat: Decompose profile button into focused components  - Create ProfileHeader component for user information display  - Create MenuItem component for reusable menu items  - Create MenuItems component for menu navigation management  - Create LogoutButton component for logout functionality  - Create ProfileMenu component for dropdown menu orchestration  - Create main ProfileButton component as orchestrator  - Update original file to re-export new component structure  - Fix TypeScript ref type mismatch in ProfileMenu component
